Move att-storage to be placed in util
