+++ examples/all-clusters-minimal-app/linux/main-common.cpp
@@ -19,7 +19,7 @@
 #include "include/tv-callbacks.h"
 #include <app-common/zap-generated/attributes/Accessors.h>
 #include <app/CommandHandler.h>
-#include <app/att-storage.h>
+#include <app/util/att-storage.h>
 #include <app/clusters/identify-server/identify-server.h>
 #include <app/server/Server.h>
 #include <app/util/af.h>
