+++ examples/chef/esp32/main/main.cpp
@@ -40,7 +40,7 @@
 
 #include <app-common/zap-generated/callback.h>
 #include <app-common/zap-generated/cluster-objects.h>
-#include <app/att-storage.h>
+#include <app/util/att-storage.h>
 #include <app/server/Dnssd.h>
 #include <app/util/af.h>
 #include <setup_payload/QRCodeSetupPayloadGenerator.h>
