+++ src/app/AttributePathExpandIterator.cpp
@@ -22,7 +22,7 @@
 #include <app/ConcreteAttributePath.h>
 #include <app/EventManagement.h>
 #include <app/GlobalAttributes.h>
-#include <app/att-storage.h>
+#include <app/util/att-storage.h>
 #include <lib/core/CHIPCore.h>
 #include <lib/core/TLVDebug.h>
 #include <lib/support/CodeUtils.h>
