+++ src/app/clusters/group-key-mgmt-server/group-key-mgmt-server.cpp
@@ -20,7 +20,7 @@
 #include <app/AttributeAccessInterface.h>
 #include <app/CommandHandler.h>
 #include <app/MessageDef/StatusIB.h>
-#include <app/att-storage.h>
+#include <app/util/att-storage.h>
 #include <app/server/Server.h>
 #include <app/util/af.h>
 #include <app/util/attribute-storage.h>
