+++ src/app/clusters/groups-server/groups-server.cpp
@@ -21,7 +21,7 @@
 #include <app-common/zap-generated/cluster-objects.h>
 #include <app-common/zap-generated/ids/Clusters.h>
 #include <app/CommandHandler.h>
-#include <app/att-storage.h>
+#include <app/util/att-storage.h>
 #include <app/reporting/reporting.h>
 #include <app/util/af.h>
 #include <app/util/config.h>
