+++ src/app/clusters/mode-select-server/mode-select-server.cpp
@@ -22,7 +22,7 @@
 #include <app/AttributeAccessInterface.h>
 #include <app/CommandHandler.h>
 #include <app/ConcreteCommandPath.h>
-#include <app/att-storage.h>
+#include <app/util/att-storage.h>
 #include <app/clusters/mode-select-server/supported-modes-manager.h>
 #include <app/util/af.h>
 #include <app/util/attribute-storage.h>
