+++ src/app/util/attribute-storage.h
@@ -25,7 +25,7 @@
 #include <app/util/endpoint-config-api.h>
 #include <lib/support/CodeUtils.h>
 
-#include <app/att-storage.h>
+#include <app/util/att-storage.h>
 #include <zap-generated/endpoint_config.h>
 
 #include <protocols/interaction_model/StatusCode.h>
