+++ src/app/util/ember-compatibility-functions.cpp
@@ -28,7 +28,7 @@
 #include <app/GlobalAttributes.h>
 #include <app/InteractionModelEngine.h>
 #include <app/RequiredPrivilege.h>
-#include <app/att-storage.h>
+#include <app/util/att-storage.h>
 #include <app/reporting/Engine.h>
 #include <app/reporting/reporting.h>
 #include <app/util/af.h>
