+++ src/darwin/Framework/CHIP/ServerEndpoint/MTRIMDispatch.mm
@@ -18,7 +18,7 @@
 #include <app-common/zap-generated/cluster-objects.h>
 #include <app/CommandHandler.h>
 #include <app/ConcreteCommandPath.h>
-#include <app/att-storage.h>
+#include <app/util/att-storage.h>
 
 #include <app/util/af-types.h>
 #include <app/util/privilege-storage.h>
