+++ src/darwin/Framework/CHIP/ServerEndpoint/MTRServerEndpoint.mm
@@ -30,7 +30,7 @@
 // TODO: These af-types.h and att-storage.h and attribute-storage.h and
 // endpoint-config-api.h and probably CodeUtils.h bits are a hack that should
 // eventually go away.
-#include <app/att-storage.h>
+#include <app/util/att-storage.h>
 #include <app/util/af-types.h>
 #include <app/util/attribute-storage.h>
 #include <app/util/endpoint-config-api.h>
